+++ libraries/ml/src/macromolecule-distance-functions/types.ts
@@ -2,4 +2,5 @@ export type mmDistanceFunctionType = (seq1: string, seq2: string) => number;
 export type mmDistanceFunctionArgs = {
     scoringMatrix: number[][];
     alphabetIndexes: {[id:string]:number};
+    threshold?: number;
 };