+++ packages/ApiSamples/scripts/ui/viewers/trellis-plot-with-summary-viewer.js
@@ -1,0 +1,14 @@
+const df = grok.data.demo.demog();
+const tv = grok.shell.addTableView(df);
+
+tv.addViewer(DG.VIEWER.TRELLIS_PLOT, {
+  viewerType: 'Summary',
+  innerViewerLook: {
+    columnNames: ['weight', 'height'],
+    aggregations: [DG.STATS.MIN, DG.STATS.MED],
+    visualizationType: 'bars',
+    colorColumnName: 'age',
+    colorAggrType: 'stdev',
+    invertColorScheme: true
+  }
+});