+++ packages/Bio/src/tests/activity-cliffs-utils.ts
@@ -5,12 +5,13 @@ import {expect} from '@datagrok-libraries/utils/src/test';
 import {activityCliffs, BYPASS_LARGE_DATA_WARNING} from '../package';
 import {DimReductionMethods} from '@datagrok-libraries/ml/src/reduce-dimensionality';
 
-export async function _testActivityCliffsOpen(df: DG.DataFrame, numberCliffs: number, method: DimReductionMethods,
-  colName: string) {
+export async function _testActivityCliffsOpen(df: DG.DataFrame, drMethod: DimReductionMethods,
+  seqColName: string, activityColName: string, similarityThr: number, tgtNumberCliffs: number
+): Promise<void> {
   await grok.data.detectSemanticTypes(df);
   const scatterPlot = await activityCliffs(
-    df, df.getCol(colName), df.getCol('activity'),
-    90, method, {[`${BYPASS_LARGE_DATA_WARNING}`]: true});
+    df, df.getCol(seqColName), df.getCol(activityColName),
+    similarityThr, drMethod, {[`${BYPASS_LARGE_DATA_WARNING}`]: true});
   // const scatterPlot = (await grok.functions.call('Bio:activityCliffs', {
   //   table: df, molecules: df.getCol(colName), activities: df.getCol('Activity'),
   //   similarity: 50, methodName: method
@@ -27,5 +28,5 @@ export async function _testActivityCliffsOpen(df: DG.DataFrame, numberCliffs: nu
     const classList: string[] = el.className.split(' ');
     return ['ui-btn', 'ui-btn-ok'].every((reqClassName) => classList.includes(reqClassName));
   });
-  expect((cliffsLink as HTMLElement).innerText.toLowerCase(), `${numberCliffs} cliffs`);
+  expect((cliffsLink as HTMLElement).innerText.toLowerCase(), `${tgtNumberCliffs} cliffs`);
 }