+++ libraries/ml/CHANGELOG.md
@@ -1,5 +1,11 @@
 # ml changelog
 
+## 6.4.0 (WIP)
+
+### Bug fixes
+
+* Fix adding type SequenceSpaceFunc
+
 ## 6.3.39 (2023-07-21)
 
 This release focuses on improving usability.